+++ qupath-extension-processing/src/main/java/qupath/process/gui/commands/density/DensityMapUI.java
@@ -307,10 +307,15 @@ public class DensityMapUI {
 	}
 
 	/**
-	 * Ignore classification (accept all objects).
+	 * Ignore classification entirely (i.e. accept all objects).
 	 * Generated with a UUID for uniqueness, and because it should not be serialized.
 	 */
-	public static final PathClass ANY_CLASS = PathClass.fromString(UUID.randomUUID().toString());
+	public static final PathClass ANY_CLASS_OR_NONE = PathClass.fromString(UUID.randomUUID().toString());
+
+	/**
+	 * Any classification, but not no classification.
+	 */
+	public static final PathClass ANY_SPECIFIED_CLASS = PathClass.fromString(UUID.randomUUID().toString());
 
 	/**
 	 * Accept any positive classification, including 1+, 2+, 3+.
@@ -887,7 +892,7 @@ public class DensityMapUI {
 			var dialog = new Dialog<ButtonType>();
 			dialog.setTitle(title);
 			dialog.setHeaderText("How do you want to export the density map?");
-			dialog.setContentText("Choose 'Raw values' of 'Send to ImageJ' if you need the original counts, or 'Color overlay' if you want to keep the same visual appearance.");
+			dialog.setContentText("Choose 'Raw values' or 'Send to ImageJ' if you need the original counts, or 'Color overlay' if you want to keep the same visual appearance.");
 			var btOrig = new ButtonType("Raw values");
 			var btColor = new ButtonType("Color overlay");
 			var btImageJ = new ButtonType("Send to ImageJ");